moves listener functions into separate directory
