+++ client/lib/listeners/clipboard.js
