+++ client/lib/listeners/delete-filters-facets.js
