+++ client/lib/listeners/display-facet.js
