+++ client/lib/listeners/display-filters.js
