+++ client/lib/listeners/download-image.js
