+++ client/lib/listeners/expand-details-button.js
