+++ client/lib/listeners/init-components.js
@@ -1,9 +1,9 @@
 var svg4everybody = require('svg4everybody');
-var searchBox = require('../lib/search-box');
+var searchBox = require('./search-box');
 var clipboard = require('./clipboard');
 var moreButton = require('./more-button');
 var carousel = require('./carousel');
-var moreDetails = require('./expand-details-button.js');
+var moreDetails = require('./expand-details-button');
 
 module.exports = (ctx) => {
   svg4everybody();
