+++ client/lib/listeners/search-listener.js
@@ -2,9 +2,10 @@
 * Add event listener to the searchbox
 */
 var page = require('page');
-var loadingBar = require('./loading-bar');
-var paramify = require('../../lib/helpers/paramify.js');
-var querify = require('../../lib/helpers/querify.js');
+
+var loadingBar = require('../loading-bar');
+var paramify = require('../../../lib/helpers/paramify.js');
+var querify = require('../../../lib/helpers/querify.js');
 
 module.exports = function () {
   var searchBoxEl = document.getElementById('searchbox');
