+++ client/routes/resources.js
@@ -1,14 +1,18 @@
-var getData = require('../lib/get-data.js');
-var JSONToHTML = require('../../lib/transforms/json-to-html-data');
 var Snackbar = require('snackbarlightjs');
-var initComp = require('../lib/init-components.js');
-var searchListener = require('../lib/search-listener');
-var osdListener = require('../lib/osd-listener');
-var downloadImageListener = require('../lib/download-image');
-var archiveListeners = require('../lib/archive-listeners');
-var getArticles = require('../lib/get-articles');
-var hideKeyboard = require('../lib/hide-keyboard');
+
 var Templates = require('../templates');
+
+var JSONToHTML = require('../../lib/transforms/json-to-html-data');
+
+var getData = require('../lib/get-data.js');
+var hideKeyboard = require('../lib/hide-keyboard');
+
+var getArticles = require('../lib/listeners/get-articles');
+var searchListener = require('../lib/listeners/search-listener');
+var downloadImageListener = require('../lib/listeners/download-image');
+var osdListener = require('../lib/listeners/osd-listener');
+var archiveListeners = require('../lib/listeners/archive-listeners');
+var initComp = require('../lib/listeners/init-components.js');
 
 module.exports = function (type) {
   return {
